+++ src/Assets/inc/CheckBox.hpp
@@ -121,7 +121,7 @@ public:
      * 
      * @param newLocation The new location of the component
      */
-    virtual void updateLocation(const Point&);
+    virtual void updateLocation( const Point& );
 
     // ----- Getters -----
 
@@ -132,6 +132,15 @@ public:
      * @return false 
      */
     bool isChecked() const;
+
+    /**
+     * @brief Scales a component 
+     * 
+     * @param factorX The X factor
+     * @param factorY The Y factor
+     */
+    void scale( const float&, const float& ) {}
+
 };
 
 }