+++ src/Assets/inc/Label.hpp
@@ -50,6 +50,8 @@ private:
     ::sf::Text _text;
     ::sf::Color _textColor;
 
+    unsigned _standardCharSize;
+
     virtual void draw(::sf::RenderTarget&, ::sf::RenderStates) const;
 
     // ----- Initializers -----
@@ -114,7 +116,16 @@ public:
      * 
      * @param newLocation The new location of the component
      */
-    virtual void updateLocation(const Point&);
+    virtual void updateLocation( const Point& );
+
+    /**
+     * @brief Scales a component 
+     * 
+     * @param factorX The X factor
+     * @param factorY The Y factor
+     */
+    void scale( const float&, const float& );
+
 };
 
 }