+++ src/Assets/inc/Textbox.hpp
@@ -177,7 +177,16 @@ public:
      * 
      * @param newLocation The new location of the component
      */
-    virtual void updateLocation(const Point&);
+    virtual void updateLocation( const Point& );
+
+    /**
+     * @brief Scales a component 
+     * 
+     * @param factorX The X factor
+     * @param factorY The Y factor
+     */
+    void scale( const float&, const float& ) {}
+
 };
 
 }