+++ src/Assets/src/Converter.cpp
@@ -64,4 +64,13 @@ CheckBox* Converter::getCheckBox(Component* asset)
     return dynamic_cast<CheckBox*>(asset);
 }
 
+Image* Converter::getImage(Component* asset)
+{
+    if(asset == nullptr)
+        return nullptr;
+
+    return dynamic_cast<Image*>(asset);
+}
+
+
 }